% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR.R
\name{BNPR}
\alias{BNPR}
\alias{BNPR_PS}
\alias{BNPR_PS_with_RD}
\title{Bayesian nonparametric phylodynamic reconstruction.}
\usage{
BNPR(
  data,
  lengthout = 100,
  pref = FALSE,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_mean = 0,
  beta1_prec = 0.001,
  rd_prob_fn = NULL,
  fns = NULL,
  log_fns = TRUE,
  fns_coeff_prior_mean = NULL,
  fns_coeff_prior_prec = NULL,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE,
  link = 1
)

BNPR_PS(
  data,
  lengthout = 100,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_mean = 0,
  beta1_prec = 0.001,
  rd_prob_fn = NULL,
  fns = NULL,
  log_fns = TRUE,
  fns_coeff_prior_mean = NULL,
  fns_coeff_prior_prec = NULL,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE,
  link = 1
)

BNPR_PS_with_RD(
  data,
  historic_sample_time,
  historic_report_time,
  max_time,
  rd_as_offset = TRUE,
  lengthout = 100,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_mean = 0,
  beta1_prec = 0.001,
  fns = NULL,
  log_fns = TRUE,
  fns_coeff_prior_mean = NULL,
  fns_coeff_prior_prec = NULL,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE,
  link = 1
)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}.}

\item{lengthout}{numeric specifying number of grid points.}

\item{pref}{logical. Should the preferential sampling model be used? If so use BNPR_PS for a better report of estimation}

\item{prec_alpha}{numeric; hyperparameter alpha for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{prec_beta}{numeric; hyperparameter beta for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{beta1_mean}{numeric; mean of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{beta1_prec}{numeric; precision of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{rd_prob_fn}{function; a function that takes in a vector of sampling times and returns the probability of a collected sample having been reported}

\item{fns}{function; list of covariate functions for the sampling intensity}

\item{log_fns}{Boolean; specifies if the log of the covariate functions, fns, needs to be take. FALSE indicates that the covriate function already returns log transformed values}

\item{fns_coeff_prior_mean}{value to specify mean for normal prior for fn coefficient}

\item{fns_coeff_prior_prec}{value to specify precision for normal prior for fn coefficient}

\item{simplify}{logical whether to fully bucket all Poisson points.}

\item{derivative}{logical whether to calculate estimates of the log-derivative.}

\item{forward}{logical whether to use the finite difference approximations of
the log-derivative as a forward or backward derivative.}

\item{link}{link for INLA "regression"}

\item{historic_sample_time}{numeric vector with historic times samples were collected (GISAID data base can be useful for obtaining such data for a given location)}

\item{historic_report_time}{numeric vector with historic times sequenced samples were reported (GISAID data base can be useful for obtaining such data for a given location)}

\item{max_time}{numeric; oldest sampling time in data}

\item{rd_as_offset}{logical whether reporting delay adjustment should be implememented as offset (TRUE), or as a covariate function with a steep prior (FALSE)}
}
\value{
Phylodynamic reconstruction of effective population size at grid points:\describe{ 
  \item{result}{contains the INLA output}
  \item{data}{contains the information passed to INLA}
  \item{grid}{contains the grid end points}
  \item{x}{contains the grid point centers} 
  \item{effpop}{contains a vector of the posterior median effective population size estimates}
  \item{effpop025}{The 2.5th posterior percentiles}
  \item{effpop975}{The 97.5th posterior percentiles} 
  \item{summary}{contains a data.frame of the estimates}
  \item{derivative}{(if \code{derivative = TRUE}) contains a data frame summarizing the log-derivative}
 }
}
\description{
Bayesian nonparametric phylodynamic reconstruction.
}
\section{Functions}{
\itemize{
\item \code{BNPR_PS}: Uses preferential sampling model.

\item \code{BNPR_PS_with_RD}: Uses preferential sampling model with adjustment for reporting delay
}}

