% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_PS_with_RD.R
\name{BNPR_PS_with_RD}
\alias{BNPR_PS_with_RD}
\title{Bayesian nonparametric phylodynamic reconstruction with reporting delay in preferential sampling model}
\usage{
BNPR_PS_with_RD(
  data,
  historic_reporting_delays,
  rd_as_offset = TRUE,
  time_offset = NULL,
  lengthout = 100,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_mean = 0,
  beta1_prec = 0.001,
  fns = NULL,
  log_fns = TRUE,
  fns_coeff_prior_mean = NULL,
  fns_coeff_prior_prec = NULL,
  simplify = TRUE,
  derivative = FALSE,
  forward = TRUE,
  link = 1
)
}
\arguments{
\item{data}{\code{phylo} object or list containing vectors of coalescent 
times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}.}

\item{historic_reporting_delays}{numeric vector with historic reporting delays. This must be on in the same units as time grid with all nonnegative values. (sequencing data bases can be useful for obtaining such data for a given location)}

\item{rd_as_offset}{logical whether reporting delay adjustment should be implemented as offset (TRUE), or as a covariate function with a steep prior (FALSE)}

\item{time_offset}{numeric; For BNPR_PS_with_RD only. The time between the first observed sampling time and time zero of analysis. Time zero is considered to be the first sampling time for BNPR() and BNPR_PS(), but BNPR_PS_with_RD() recognizes that this will likely not be true. This argument shifts the time to correspond to time zero of analysis, which is both helpful practically, and necessary for the reporting probability function to correctly assign probability or being reported based on sampling time.}

\item{lengthout}{numeric specifying number of grid points.}

\item{prec_alpha}{numeric; hyperparameter alpha for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{prec_beta}{numeric; hyperparameter beta for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{beta1_mean}{numeric; mean of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{beta1_prec}{numeric; precision of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{fns}{function; list of covariate functions for the sampling intensity}

\item{log_fns}{logical; specifies if the log of the covariate functions, fns, needs to be taken. FALSE indicates that the covariate function already returns log transformed values}

\item{fns_coeff_prior_mean}{numeric vector; normal prior mean for fns coefficient(s). If non NULL, must match length of fns}

\item{fns_coeff_prior_prec}{numeric vector; normal prior precision for fns coefficient(s). If non NULL, must match length of fns}

\item{simplify}{logical whether to fully bucket all Poisson points.}

\item{derivative}{logical whether to calculate estimates of the log-derivative.}

\item{forward}{logical whether to use the finite difference approximations of
the log-derivative as a forward or backward derivative.}

\item{link}{link for INLA "regression"}
}
\value{
Phylodynamic reconstruction of effective population size at grid points:\describe{ 
  \item{result}{contains the INLA output}
  \item{data}{contains the information passed to INLA}
  \item{grid}{contains the grid end points}
  \item{x}{contains the grid point centers} 
  \item{effpop}{contains a vector of the posterior median effective population size estimates}
  \item{effpop025}{The 2.5th posterior percentiles}
  \item{effpop975}{The 97.5th posterior percentiles} 
  \item{summary}{contains a data.frame of the estimates}
  \item{derivative}{(if \code{derivative = TRUE}) contains a data frame summarizing the log-derivative}
  \item{rd_prob_fn}{Function which computes probability of a sample having been reported based on it's sampling time. Will be log probabilities unless reporting delay function was specified as covariate and log_fns is FALSE.}
 }
}
\description{
Bayesian nonparametric phylodynamic reconstruction with reporting delay in preferential sampling model
}
