% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_phylo.R
\name{branching_sampling_times}
\alias{branching_sampling_times}
\title{branching_sampling_times Title TODO}
\usage{
branching_sampling_times(tr)
}
\arguments{
\item{tr}{a \code{phylo} object containing a phylogeny}
}
\value{
numeric vector TODO
}
\description{
branching_sampling_times Title TODO
}
