% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent-model-helpers.R
\name{coal_stats}
\alias{coal_stats}
\title{coal_stats Returns data frame on midpoints of regular grid with record of coalescent events and Poisson regression offset}
\usage{
coal_stats(grid, samp_times, coal_times, n_sampled = NULL, log_zero = -100)
}
\arguments{
\item{grid}{numeric vector; regular grid over entire genealogy}

\item{samp_times}{numeric vector; sampling times}

\item{coal_times}{numeric vector; coalescence times}

\item{n_sampled}{numeric vector; number of sampling events at each sampling time, length matching the length of samp_times}

\item{log_zero}{numeric; value to store for E of log(0)}
}
\value{
a data frame with \describe{
  \item{time}{Vector of midpoints of regular grid where each value represents that either a grid end point, sampling time, or coalescent time occured in that regular grid interval}
  \item{event}{Logical vector recording coalescent events with a 1 and regurlar grid points or sampling times with a 0}
  \item{E}{Coalescent factor times width of interval (on combined grid level)}
  \item{E_log}{Possion regression offset}
}
}
\description{
coal_stats Returns data frame on midpoints of regular grid with record of coalescent events and Poisson regression offset
}
