% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalsim.R
\name{coalsim_thin}
\alias{coalsim_thin}
\title{Coalescent thinning method}
\usage{
coalsim_thin(samp_times, n_sampled, traj, lower_bound, ...)
}
\arguments{
\item{samp_times}{numeric vector of sampling times}

\item{n_sampled}{numeric vector of number of samples collected at each sampling time}

\item{traj}{function that returns effective population size at time t.}

\item{lower_bound}{numeric lower limit of \code{traj} function on its
support.  Used only by thinning method.}

\item{...}{additional arguments to be passed to \code{traj} function}
}
\value{
A list containing vectors of coalescent times \code{coal_times}, 
  intercoalescent times \code{intercoal_times}, and number of active lineages
  \code{lineages}, as well as passing along \code{samp_times} and
  \code{n_sampled}.
}
\description{
Coalescent thinning method
}
