% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal_samp.R
\name{condense_stats}
\alias{condense_stats}
\title{condense_stats Title TODO}
\usage{
condense_stats(time, event, E, log_zero = -100)
}
\arguments{
\item{time}{numeric vector; time of each event}

\item{event}{numeric vector; number of events at a given time}

\item{E}{numeric vector; TODO}

\item{log_zero}{numeric; value to store for E of log(0)}
}
\value{
data frame with \describe{
  \item{time}{column of times}
  \item{event}{number of events at each time}
  \item{E}{TODO}
  \item{E_log}{TODO}
}
}
\description{
condense_stats Title TODO
}
