% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_phylo.R
\name{connect_sample_to_tips}
\alias{connect_sample_to_tips}
\title{Summarizes phylogeny and connects sampling times to tips}
\usage{
connect_sample_to_tips(phy, backwards = TRUE)
}
\arguments{
\item{phy}{\code{phylo} object containing a phylogeny}

\item{backwards}{Boolean; TODO}
}
\value{
List with: \describe{
  \item{ungrouped_samp_df}{data frame connecting sampling times to corresponding tip labels}
  \item{samp_times}{sampling times}
  \item{n_sampled}{number sampled at each sampling time}
  \item{coal_times}{coalescence times}
}
}
\description{
Summarizes phylogeny and connects sampling times to tips
}
