% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent-model-helpers.R
\name{gen_INLA_args}
\alias{gen_INLA_args}
\title{gen_INLA_args Returns data frame on combined grid with record of coalescent events and coalescent factor}
\usage{
gen_INLA_args(samp_times, n_sampled, coal_times)
}
\arguments{
\item{samp_times}{numeric vector; sampling times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{coal_times}{numeric vector; coalescence times}
}
\value{
data frame with \describe{
  \item{coal_factor}{Number of active lineages choose 2}
  \item{s}{Sampling and coalescence times combined and sorted} 
  \item{event}{Logical vector with 0 = sampling time and 1 = coalescent time}
  \item{lineages}{Numeric vector of number of active lineages}
}
}
\description{
gen_INLA_args Returns data frame on combined grid with record of coalescent events and coalescent factor
}
