% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_newick.R
\name{generate_newick}
\alias{generate_newick}
\title{Generate Newick}
\usage{
generate_newick(args)
}
\arguments{
\item{args}{is a list containing vectors of coalescent times \code{coal_times}, sampling times \code{samp_times}, and number 
sampled per sampling time \code{n_sampled}, etc. This list is the output of coalsim}
}
\value{
A list with two elements \code{newick} contains the tree in phylo format, \code{labels} a vector with tip labels
}
\description{
Creates a phylo object from the results of the \code{coalsim()} function with tip labels
}
\examples{
constant <- function(x){return (rep(1,length(x)))}
simulation1 <- coalsim(samp_times = 0, n_sampled = 10, traj = constant)
tree1 <- generate_newick(simulation1)
plot(tree1$newick)
}
