% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reported_prob.R
\name{get_reported_prob}
\alias{get_reported_prob}
\title{Returns probability of a collected sample having been reported by the analysis's time zero}
\usage{
get_reported_prob(
  sampling_times,
  historic_sample_time,
  historic_report_time,
  lengthout
)
}
\arguments{
\item{sampling_times}{numeric vector; sampling times}

\item{historic_sample_time}{numeric vector with historic times samples were collected (GISAID data base can be useful for obtaining such data for a given location)}

\item{historic_report_time}{numeric vector with historic times sequenced samples were reported (GISAID data base can be useful for obtaining such data for a given location)}

\item{lengthout}{numeric; length of time grid which defines granularity of reported probabilities, same length as using in BNPR estimation}
}
\value{
vector of probabilities of a sample having been reported by the analysis's time zero; given the sampling time
}
\description{
Returns probability of a collected sample having been reported by the analysis's time zero
}
