% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalsim.R
\name{hazard_uniroot_stepfun}
\alias{hazard_uniroot_stepfun}
\title{Hazard uniroot step function}
\usage{
hazard_uniroot_stepfun(traj_inv_stepfun, lineages, start, target)
}
\arguments{
\item{traj_inv_stepfun}{inverse of the function that returns effective population size at time t}

\item{lineages}{number of active lineages}

\item{start}{time to start at}

\item{target}{TODO}
}
\value{
result vector minus start
}
\description{
Hazard uniroot step function
}
