% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_phylo.R
\name{heterochronous_gp_stat}
\alias{heterochronous_gp_stat}
\title{heterochronous_gp_stat Title TODO}
\usage{
heterochronous_gp_stat(phy, tol = 0)
}
\arguments{
\item{phy}{a \code{phylo} object containing a phylogeny}

\item{tol}{numeric; tolerance}
}
\value{
list with \describe{
  \item{coal_times}{numeric vector; coalescent times}
  \item{samp_times}{numeric vector; sampling times}
  \item{n_sampled}{numeric vector; number sampled at each sampling time}
}
}
\description{
heterochronous_gp_stat Title TODO
}
