% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal.R
\name{infer_coal}
\alias{infer_coal}
\title{Infer coalescence}
\usage{
infer_coal(
  samp_times,
  coal_times,
  n_sampled = NULL,
  lengthout = 100,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  simplify = FALSE,
  derivative = FALSE
)
}
\arguments{
\item{samp_times}{numeric vector; sampling times}

\item{coal_times}{numeric vector; coalescence times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{lengthout}{numeric; length of time for the estimation}

\item{prec_alpha}{numeric; hyperparameter alpha for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{prec_beta}{numeric; hyperparameter beta for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{simplify}{Boolean; TODO}

\item{derivative}{Boolean; TODO}
}
\value{
list with \describe{
  \item{result}{INLA call return}
  \item{data}{data frame; data fed to the INLA call}
  \item{grid}{vector; sequence from minimum sample time to maximum sample time, of length one more than lengthout.}
  \item{x}{vector; coal data time}
}
}
\description{
Infer coalescence
}
