% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalsim.R
\name{integrate_step_fun}
\alias{integrate_step_fun}
\title{Integrates step function}
\usage{
integrate_step_fun(fun, from, to)
}
\arguments{
\item{fun}{function}

\item{from}{start}

\item{to}{end}
}
\value{
numeric; integration value
}
\description{
Integrates step function
}
