% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferential-sampling-model-helpers.R
\name{joint_stats}
\alias{joint_stats}
\title{joint_stats Prepares data for INLA for case of preferential sampling model}
\usage{
joint_stats(coal_data, samp_data)
}
\arguments{
\item{coal_data}{data frame with columns: time, event, E_log}

\item{samp_data}{data frame with columns: time, count, E_log}
}
\value{
list with \describe{
  \item{Y}{matrix with two columns. The first column contains the coalescence events padded with NA's at end. The second column contains the sampling counts, padded by NA's at the beginning.}
  \item{beta0}{numeric vector with 0's for length of coal_data$time and 1's for length of samp_data$time; for sampling intensity function}
  \item{time}{coalescence times padded with 0's at end}
  \item{time2}{sampling times padded with 0's at beginning}
  \item{w}{numeric vector; 1's repeated for length of coal_data$time followed by -1's repeated for length of samp_data$time}
  \item{E_log}{coalescence E_log followed by sampling E_log}
}
}
\description{
joint_stats Prepares data for INLA for case of preferential sampling model
}
