% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferential-sampling-model-helpers.R
\name{samp_stats}
\alias{samp_stats}
\title{samp_stats Collect sampling data for estimation accounting for preferential sampling,}
\usage{
samp_stats(grid, samp_times, n_sampled = NULL, trim_end = FALSE)
}
\arguments{
\item{grid}{numeric vector; regular grid across genealogy}

\item{samp_times}{numeric vector; sampling times}

\item{n_sampled}{numeric vector; number of sampling events at each sampling time, length matching the length of samp_times}

\item{trim_end}{logical; TODO}
}
\value{
data frame with \describe{
  \item{time}{Midpoints of time grid intervals}
  \item{count}{Number of sampling times within each grid interval}
  \item{E}{Vector of differences between consecutive grid points}
  \item{E_log}{Log of E}
}
}
\description{
samp_stats Collect sampling data for estimation accounting for preferential sampling,
}
