% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BNPR.R
\name{shade_band}
\alias{shade_band}
\title{Plot shade bands}
\usage{
shade_band(x, ylo, yhi, xlim = NULL, col = "gray")
}
\arguments{
\item{x}{x values corresponding to the band bounds}

\item{ylo}{lower band bound}

\item{yhi}{upper band bound}

\item{xlim}{x plotting limits}

\item{col}{color}
}
\description{
Plot shade bands
}
