% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_phylo.R
\name{summarize_phylo2}
\alias{summarize_phylo2}
\title{Summarizes phylogeny, connecting sampling times to tips}
\usage{
summarize_phylo2(phy, backwards = TRUE)
}
\arguments{
\item{phy}{\code{phylo} object containing a phylogeny}

\item{backwards}{Boolean; TODO}
}
\value{
list with \describe{
  \item{ungrouped_samp_df}{data frame connecting sampling times to corresponding tip labels}
  \item{samp_times}{sampling times}
  \item{n_sampled}{number sampled at each sampling time}
  \item{coal_times}{coalescence times}
}
}
\description{
Summarizes phylogeny, connecting sampling times to tips
}
