% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_tree.R
\name{trim_tree_with_delays}
\alias{trim_tree_with_delays}
\title{Title}
\usage{
trim_tree_with_delays(full_tree, calc_report_prob, original_samp_times)
}
\arguments{
\item{full_tree}{`phylo` object containing a phylogeny}
}
\description{
Title
}
