% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washington_delays}
\alias{washington_delays}
\title{Washington SARS-CoV-2 Sampling and Reporting Dates}
\format{
## `washington_delays`
A data frame with 56 rows and 2 columns:
\describe{
  \item{sampling_date}{Date the sample was collected.}
  \item{reporting_date}{Date the sample's genetic sequence was reported to the database.}
}
}
\source{
<https://doi.org/10.55876/gis8.240619zd>
}
\usage{
washington_delays
}
\description{
Sampling and reporting dates for a random subset of SARS-CoV-2 sequences in 
the state of Washington reported in July of 2021 to GISAID.
}
\keyword{datasets}
