% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washington_genealogy}
\alias{washington_genealogy}
\title{Washington SARS-CoV-2 Genealogy}
\format{
An object of class \code{phylo} of length 4.
}
\source{
<https://github.com/CatalinaMedina/reporting-delays-in-phylodynamics-paper>
}
\usage{
washington_genealogy
}
\description{
This dataset is a \code{phylo} object of a genealogy generated by BEAST from
the sequence data in Medina, et al. 2024.
}
\details{
C. M. Medina, J. A. Palacios, V. M. Minin (2024). Accounting for reporting 
delays in real-time phylodynamic analyses with preferential sampling. 
TODO: Update with paper location
}
\keyword{datasets}
