% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal.R
\name{coal_stats}
\alias{coal_stats}
\title{coal_stats Title TODO}
\usage{
coal_stats(grid, samp_times, coal_times, n_sampled = NULL, log_zero = -100)
}
\arguments{
\item{grid}{numeric vector; TODO}

\item{samp_times}{numeric vector; sampling times}

\item{coal_times}{numeric vector; coalescence times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{log_zero}{numeric; value to store for E of log(0)}
}
\value{
a data frame with \describe{
  \item{time}{TODO}
  \item{event}{TODO}
  \item{E}{TODO}
  \item{E_log}{TODO}
}
}
\description{
coal_stats Title TODO
}
