% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent-model-helpers.R
\name{condense_stats}
\alias{condense_stats}
\title{condense_stats Sums all Poisson regression offsets (product of coalescent and combined interval width) within each regular grid interval}
\usage{
condense_stats(time, event, E, log_zero = -100)
}
\arguments{
\item{time}{numeric vector; time of each event}

\item{event}{numeric vector; number of events at a given time}

\item{E}{numeric vector; products of coalescent factor and width of interval (on combined grid level)}

\item{log_zero}{numeric; value to store for E of log(0)}
}
\value{
data frame with \describe{
  \item{time}{Vector of midpoints of regular grid where each value represents that either a grid end point, sampling time, or coalescent time occured in that regular grid interval}
  \item{event}{Number of coalescent events within each regular grid interval}
  \item{E}{Vector of the sum of products of coalescent factor and width of interval (on combined grid level), for each regular grid interval}
  \item{E_log}{Possion regression offset}
}
}
\description{
condense_stats Sums all Poisson regression offsets (product of coalescent and combined interval width) within each regular grid interval
}
