% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal.R
\name{gen_INLA_args}
\alias{gen_INLA_args}
\title{gen_INLA_args Title TODO}
\usage{
gen_INLA_args(samp_times, n_sampled, coal_times)
}
\arguments{
\item{samp_times}{numeric vector; sampling times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{coal_times}{numeric vector; coalescence times}
}
\value{
data frame with \describe{
  \item{coal_factor}{TODO}
  \item{s}{sampling and coalescence times combined and sorted} 
  \item{event}{TODO}
  \item{lineages}{TODO}
}
}
\description{
gen_INLA_args Title TODO
}
