% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal_samp.R
\name{infer_coal_samp}
\alias{infer_coal_samp}
\title{Infer coalescence sample}
\usage{
infer_coal_samp(
  samp_times,
  coal_times,
  n_sampled = NULL,
  lengthout = 100,
  rd_prob_fn = NULL,
  fns = NULL,
  log_fns = TRUE,
  fns_coeff_prior_mean = NULL,
  fns_coeff_prior_prec = NULL,
  prec_alpha = 0.01,
  prec_beta = 0.01,
  beta1_mean = 0,
  beta1_prec = 0.001,
  use_samp = FALSE,
  simplify = FALSE,
  events_only = FALSE,
  derivative = FALSE,
  link = 1
)
}
\arguments{
\item{samp_times}{numeric vector; sampling times}

\item{coal_times}{numeric vector; coalescence times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{lengthout}{numeric; length of time for the estimation}

\item{rd_prob_fn}{function; a function that takes in a vector of sampling times and returns the probability of a collected sample having been reported}

\item{fns}{function; list of covariate functions for the sampling intensity}

\item{log_fns}{Boolean; specifies if the log of the covariate functions, fns, needs to be take. FALSE indicates that the covriate function already returns log transformed values}

\item{fns_coeff_prior_mean}{numeric vector; normal prior mean for fns coefficient(s). If non NULL, must match length of fns}

\item{fns_coeff_prior_prec}{numeric vector; normal prior precision for fns coefficient(s). If non NULL, must match length of fns}

\item{prec_alpha}{numeric; hyperparameter alpha for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{prec_beta}{numeric; hyperparameter beta for the gamma prior of kappa, the precision of the Gaussian random walk prior}

\item{beta1_mean}{numeric; mean of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{beta1_prec}{numeric; precision of the normal prior assigned to the coefficient of the log effective population size in the sampling intensity formula}

\item{use_samp}{Boolean; TODO}

\item{simplify}{logical whether to fully bucket all Poisson points.}

\item{events_only}{Boolean; TODO}

\item{derivative}{logical whether to calculate estimates of the log-derivative.}

\item{link}{link for INLA "regression"}
}
\value{
list with \describe{
  \item{result}{INLA call return}
  \item{data}{data frame; data fed to the INLA call}
  \item{grid}{vector; sequence from minimum sample time to maximum sample time, of length one more than lengthout.}
  \item{x}{vector; coal data time}
}
}
\description{
Infer coalescence sample
}
