% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_coal.R
\name{samp_stats}
\alias{samp_stats}
\title{samp_stats Title TODO}
\usage{
samp_stats(grid, samp_times, n_sampled = NULL, trim_end = FALSE)
}
\arguments{
\item{grid}{numeric vector; TODO}

\item{samp_times}{numeric vector; sampling times}

\item{n_sampled}{numeric vector; The number of sampling events at each sampling time, length matching the length of samp_times}

\item{trim_end}{Boolean; TODO}
}
\value{
data frame with \describe{
  \item{time}{midpoints of grid}
  \item{count}{TODO}
  \item{E}{difference between consecutive grid points}
  \item{E_log}{log of E}
}
}
\description{
samp_stats Title TODO
}
