% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_tree_with_delays.R
\name{trim_tree_with_delays}
\alias{trim_tree_with_delays}
\title{Trim phylogeny with reporting delays}
\usage{
trim_tree_with_delays(full_phylo, calc_report_prob, original_samp_times)
}
\arguments{
\item{full_phylo}{a phylo object containing a phylogeny}

\item{calc_report_prob}{a function which returns probability of a sample having been reported by time zero of analysis given the time at which the sample was collected}

\item{original_samp_times}{a numeric vector of the sampling times for the \code{full_tree}}
}
\value{
A list with two elements \code{obs_phylo} contains the tree in phylo format only containing samples that were simulated as observed, \code{time0_offset} a numeric of the most recently observed sampling time. This is useful in simulations because the phylo object will automatically consider time zero to be the most recently observed sampling time.
}
\description{
Function to simulate reporting delays from provided \code{calc_report_prob()} according to sampling times. Useful for subsetting tree to only observed samples and keeping track of original time zero.
}
